more work on Login, renamed to LoginPage
